splitting const files
