+++ my-app/src/app/utils/consts/fieldNames.js
@@ -1,3 +1,2 @@
-export const PATH = '/';
 export const MAIL_FIELD_NAME = 'mail';
 export const PASSWORD_FIELD_NAME = 'password';
