+++ my-app/src/app/utils/consts/routes.js
@@ -1,0 +1,1 @@
+export const PATH = '/';
